backend/requirements/base.txt
Django>=5.1,<6.0
djangorestframework>=3.15.2,<4.0
httpx>=0.27,<0.28
Pillow>=11.0,<12.0
